Terminal "graphical" upgrades.
 - Bigger terminal virtual size.

 - Showing a fancy ASCII art with the term_showSplashScreen function.

 - Added a function in the framebuffer to handle newlines.

+++ terminal.h
@@ -44,13 +44,12 @@ void term_move_left();
 void term_move_up();
 void term_move_down();
 
-void term_set_cursor(uint8_t col, uint8_t row);
+void term_set_cursor(int col, int row);
 
 void term_write(unsigned char c);
 void term_erase();
 
 void term_enter();
-
 
 void term_showSplashScreen();
